refactor: Phase 2 - main_test.go出力関数テスト分離
大きなmain_test.goファイルから出力フォーマット関連のテスト関数を
output_test.goに分離して保守性を向上

## 変更内容

### 新規ファイル
- cmd/beaver/output_test.go: 出力フォーマットテスト専用ファイル
  - TestOutputSummarization (197行): AI要約出力フォーマットテスト
  - TestOutputBatchSummarization (240行): バッチAI要約出力テスト
  - TestOutputSummary (151行): GitHub Issue要約出力テスト

### 変更ファイル
- cmd/beaver/main_test.go: 3つの大型テスト関数を削除
  - ファイルサイズ: ~4141行 → ~3553行 (588行削減)
  - 未使用インポートを削除してコンパイルエラー解消

## 技術仕様
- 全てのテストケースを完全に保持
- 関数間の依存関係なく安全に分離
- 既存テストヘルパーとの互換性維持
- Phase 1で作成したutils_test.goとの統合

## テスト結果
- 全テストケースがPASS
- コンパイルエラー解消
- 品質チェック完了
